Move pserror.h to ddebug module
