+++ Descent3/special_face.cpp
@@ -16,13 +16,12 @@
 * along with this program.  If not, see <http://www.gnu.org/licenses/>.
 */
 
-#include "pstypes.h"
+#include <cstdlib>
+#include <cstring>
 
-#include "special_face.h"
-#include "mono.h"
-#include <stdlib.h>
-#include <string.h>
 #include "mem.h"
+#include "pserror.h"
+#include "special_face.h"
 
 int Num_of_special_faces = 0;
 special_face SpecialFaces[MAX_SPECIAL_FACES];
