+++ Descent3/terrainrender.cpp
@@ -794,16 +794,12 @@ void RenderMine(int viewer_roomnum, int flag_automap, int called_from_terrain, b
 #include "terrain.h"
 #include "grdefs.h"
 #include "3d.h"
-#include "pstypes.h"
 #include "pserror.h"
 #include "renderer.h"
 #include "gametexture.h"
 #include "descent.h"
 #include "render.h"
 #include "game.h"
-#include "ddio.h"
-#include "polymodel.h"
-#include "lighting.h"
 #include "vecmat.h"
 #include "renderobject.h"
 #include "findintersection.h"
