+++ linux/CMakeLists.txt
@@ -21,7 +21,7 @@ target_include_directories(linux PUBLIC
   >
 )
 target_link_libraries(linux PRIVATE
+  ddebug
   ddio
-  misc
   SDL2::SDL2
 )
