+++ mem/CMakeLists.txt
@@ -6,8 +6,8 @@ target_compile_definitions(mem PUBLIC
   $<$<BOOL:${ENABLE_MEM_RTL}>:MEM_USE_RTL>
 )
 target_link_libraries(mem PRIVATE
+  ddebug
   logger
-  misc
 )
 target_include_directories(mem PUBLIC
   $<BUILD_INTERFACE:
