+++ misc/CMakeLists.txt
@@ -1,20 +1,15 @@
 set(HEADERS
-  pserror.h
   psglob.h
   psrand.h
-  pstring.h)
+  pstring.h
+)
 set(CPPS
-  error.cpp
   psglob.cpp
   psrand.cpp
   pstring.cpp
 )
 
 add_library(misc STATIC ${HEADERS} ${CPPS})
-target_link_libraries(misc PRIVATE
-  ddebug
-  SDL2::SDL2
-)
 target_include_directories(misc PUBLIC
   $<BUILD_INTERFACE:
     ${PROJECT_SOURCE_DIR}/misc
