+++ renderer/HardwareInstance.cpp
@@ -16,9 +16,11 @@
 * along with this program.  If not, see <http://www.gnu.org/licenses/>.
 */
 
+#include <cstring>
+
 #include "3d.h"
 #include "HardwareInternal.h"
-#include <string.h>
+#include "pserror.h"
 
 struct InstanceContext {
   matrix m_viewMatrix;     // matrix
