+++ renderer/HardwareInternal.h
@@ -19,8 +19,6 @@
 #ifndef __HARDWARE_INTERNAL_H__
 #define __HARDWARE_INTERNAL_H__
 
-#include "pserror.h"
-
 #define MAX_POINTS_IN_POLY 100
 
 // These structs are for drawing with vertex arrays
