+++ renderer/HardwareTransforms.cpp
@@ -16,11 +16,11 @@
 * along with this program.  If not, see <http://www.gnu.org/licenses/>.
 */
 
+#include <cstring>
+
 #include "3d.h"
-#include "pserror.h"
 #include "HardwareInternal.h"
 #include "renderer.h"
-#include <string.h>
 
 // Whether or not to use T&L transforms or the pass-thru ones
 static int sUseTransformPassthru = -1;
