+++ vecmat/CMakeLists.txt
@@ -4,6 +4,7 @@ set(CPPS
 
 add_library(vecmat STATIC ${HEADERS} ${CPPS})
 target_link_libraries(vecmat PRIVATE
+  ddebug
   fix
   misc
 )
